requirements.txt
ansicon==1.89.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==3.2
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
ezenv==0.92
importlib-metadata==5.0.0
jinxed==1.2.0
jsonfield==3.1.0
Markdown==3.4.1
phonenumbers==8.12.57
psycopg2==2.9.4
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
swapper==1.3.0
tzdata==2022.5
zipp==3.9.0
